refactor: split decryption functions
